// N/client/src/module/Admin/admin.validator.tsx
export interface BannerProps {
    _id: string;
    title: string;
    subtitle: string;
    type: string;
    startAt: string; // ISO date string
    endAt: string;   // ISO date string
    isActive: boolean;
    priority: number;
    createdAt: string;
    updatedAt: string;
    __v: number;
    image: {
        public_id: string;
        secure_url: string;
    };
}
// ...
export interface BannerResponse {
    code: number;
    data: BannerProps[];
    message: string;
    options: {
        total: number;
    };
    status: string;
}